send email,merapihkan daftar buku,validasi tabel

+++ app/Http/Controllers/Admin/KategoriController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Kategori;
 use App\Models\Peminjamens;
+use App\Models\Buku;
 use Illuminate\Support\Facades\Auth;
 
 
@@ -78,8 +79,21 @@ class KategoriController extends Controller
 
     public function destroy($id)
     {
-        $kategori = Kategori::FindOrFail($id);
+        $kategori = Kategori::findOrFail($id);
+    
+        // Check if there are any books associated with this category that are borrowed
+        $bukudipinjam = Buku::where('id_kategori', $id)
+            ->whereHas('peminjamens', function($query) {
+                $query->whereIn('status_pengajuan', ['menunggu pengajuan', 'pengajuan diterima','pengembalian diterima', 'pengembalian ditolak','dikembalikan']);
+            })
+            ->exists();
+    
+        if ($bukudipinjam) {
+            return redirect()->route('kategori.index')->with('error', 'Kategori tidak bisa dihapus karena masih ada buku yang terkait.');
+        }
+    
         $kategori->delete();
         return redirect()->route('kategori.index')->with('success', 'Data berhasil dihapus');
     }
+    
 }
