+++ app/Http/Controllers/Admin/PenerbitController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Penerbit;
 use App\Models\Peminjamens;
+use App\Models\Buku;
 use Illuminate\Support\Facades\Auth;
 
 
@@ -79,6 +80,17 @@ class PenerbitController extends Controller
     public function destroy($id)
     {
         $penerbit = Penerbit::FindOrFail($id);
+
+        $bukudipinjam = Buku::where('id_penerbit', $id)
+        ->whereHas('peminjamens', function($query) {
+            $query->whereIn('status_pengajuan', ['menunggu pengajuan', 'pengajuan diterima','pengembalian diterima', 'pengembalian ditolak','dikembalikan']);
+        })
+        ->exists();
+
+    if ($bukudipinjam) {
+        return redirect()->route('penerbit.index')->with('error', 'Penerbit tidak bisa dihapus karena masih ada buku yang terkait.');
+    }
+        
         $penerbit->delete();
         return redirect()->route('penerbit.index')->with('success', 'Data berhasil dihapus');
     }
